Bugfix, refactor AI functions
